+++ lib/struct/alias_member/subclassinstancemethods.rb
@@ -1,26 +1,16 @@
 class Struct; module Alias_Member
   
-  module InstanceMethods
+  module SubclassInstanceMethods
   
-    # @param [Symbol, String, #to_sym, Fixnum] key
+    # @param [Symbol, String, #to_sym, Integer, #to_int] key
     def [](key)
-      case key
-      when Symbol, String
-        super self.class.autonym(key)
-      else
-        super key, value
-      end
+      super self.class.autonym_for_key(key)
     end
     
-    # @param [Symbol, String, #to_sym, Fixnum] key
+    # @param [Symbol, String, #to_sym, Integer, #to_int] key
     # @return value
     def []=(key, value)
-      case key
-      when Symbol, String
-        super self.class.autonym(key), value
-      else
-        super key, value
-      end
+      super self.class.autonym_for_key(key), value
     end
 
     # @see [ClassMethods#aliases]