+++ History.md
@@ -1,5 +1,13 @@
+0.0.2 2012-09-16
+================
+
+* Fix when the Struct more and more extended by anyone
+* Modify
+  before: Subclass.autonym
+  after:  Subclass.autonym_for_name, Subclass.autonym_for_key
+
 0.0.1 2012-07-09
 ================
 
-Initial release
+* Initial release